refactors code structure
