+++ brado/src/cnpj/mod.rs
@@ -1,0 +1,2 @@
+mod cnpj;
+pub use cnpj::*;
