--- brado/src/docs/mod.rs
@@ -1,5 +1,0 @@
-pub mod cnh;
-pub mod cnpj;
-pub mod common;
-pub mod cpf;
-pub mod doc;
