+++ brado/src/lib.rs
@@ -5,12 +5,10 @@
 //! ## Example
 //! ```
 //! use brado;
-//! use brado::Document;
-//! let cpf_doc: Document = Document::new("639.292.470-11");
-//! brado::cpf::validate(&cpf_doc, true, false);
+//! let cpf_doc = String::from("639.292.470-11");
+//! brado::cpf::validate(&cpf_doc, true);
 //! ```
-mod docs;
-pub use crate::docs::cnh;
-pub use crate::docs::cnpj;
-pub use crate::docs::cpf;
-pub use crate::docs::doc::Document;
+pub mod common;
+pub mod cpf;
+pub mod cnpj;
+pub mod cnh;
